pyproject.toml
[tool.poetry]
package-mode = false
name = "news-aggregator"
version = "0.1.0"
description = "news aggregator"
authors = ["n529joker <ngengerawlings004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.4"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
databases = "^0.9.0"
strawberry-graphql = "^0.237.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

